chore: Add linting for JavaScript
This adds linting for JavaScript and JavaScript related stuff, i.e.
JSX, React and JSDoc. This ensures code consistency and quality and can
potentially find errors before committing code into the repository.

There are two lint files:
* `.eslintrc.json` only produces warnings and is used during
  development, e.g. in an IDE
* `.eslintrc-commit.json` produces lint errors and is used during
  Git commits, i.e. ESLint will be executed with this set of rules
  by husky in the `pre-commit` hook. The commit will fail if there
  are lint errors.

The rule set is rather strict at the moment and may be
adjusted later on in the development process.